requirements.txt
amqp==5.2.0
Appium-Python-Client==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==3.6.4.0
celery==5.0.5
certifi==2024.2.2
chardet==4.0.0
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2
django-admin-interface==0.16.1
django-apptemplates==1.5
django-background-tasks==1.2.5
django-bootstrap4==2.3.1
django-bootstrap5==21.1
django-colorfield==0.11.0
django-compat==1.0.15
django-crontab==0.7.1
django-debug-toolbar==3.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-impersonate==1.7.3
django-jazzmin==2.4.8
django-mathfilters==1.0.0
django-ratelimit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
idna==2.10
kombu==5.3.7
monnify==1.0.0
more-itertools==10.2.0
pillow==10.3.0
prompt-toolkit==3.0.43
python-decouple==3.4
pytz==2024.1
requests==2.25.1
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
